+++ WakupDemo/WakupDemo/AppDelegate.swift
@@ -18,7 +18,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Override point for customization after application launch.
         
         WakupManager.manager.setup("1223b89c-ca0b-4c40-bc87-6fce7f7d1ad7")  // Sample project API Key, don't use this in a production environment
-        window?.rootViewController = WakupManager.manager.rootNavigationController()
+        window?.rootViewController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController()
         window?.makeKeyAndVisible()
         
         return true