+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- No changes
+### Changed
+- Some classes are now public to allow better UI customization
 
 ## [2.0.0] - 2016-10-26
 ### Added